refactor: outsource `equal*` functions for the df graph
